refactor: equipment commponet
