+++ frontend/src/components/Equipment/EquipmentCard.jsx
