+++ frontend/src/components/Equipment/EquipmentList.jsx
